requirement.txt
requests
discord.py
python-dotenv
supabase
google-genai